+++ ng/practices-ui-tailwind/jest-schematics.config.js
@@ -1,5 +1,6 @@
 module.exports = {
     ...require('./jest.config.js'),
     testEnvironment: 'node',
+    testPathIgnorePatterns: [],
+    testMatch: ['<rootDir>/schematics/**/*.spec.ts'],
 };
-